models: SVG: tree parse to support netsted groups
Instead of looking only at direct childs of a node,
search the complete tree starting at the node.

This is particularly useful when we add fancy attributes,
since they tend to add encompassing groups to the graph.

It is pretty easy to test for the bug and how it can be reproduced.
It's tracked in: https://github.com/hbmartin/graphviz2drawio/issues/45

I re-created a minimal fail case from
https://github.com/hbmartin/graphviz2drawio/issues/45#issuecomment-1122551983

```
digraph {
    Pitas
    Pizza [ tooltip=Delicious ]
    Pitas -> Pizza
}
```

Previous HEAD failed to parse this.