mx: NodeFactory: support creating blank rect from image
Image nodes are nice to have in some schematics.
Because of the representation in mxGraph, it would be a pain to create them in the xml.

Instead, this follows the least-effort methodology and replace the image by an equally sized rectangle.

In drawio GUI, the image can then be switched back in place of the rectangle.

As a reference on the mxImageShape if someone wants to dive deeper:
https://jgraph.github.io/mxgraph/java/docs/com/mxgraph/shape/mxImageShape.html